requirements.txt
absl-py==1.4.0
cachetools==5.3.0
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
cmake==3.26.0
contourpy==1.0.7
cycler==0.10.0
filelock==3.10.2
fonttools==4.39.2
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
idna==2.10
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lit==16.0.0
Markdown==3.4.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.7.0.72
pandas==1.5.3
Pillow==9.4.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
roboflow==0.2.31
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.17.0
sklearn==0.0.post1
sympy==1.11.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
triton==2.0.0
ultralytics==8.0.20
urllib3==1.26.15
Werkzeug==2.2.3
wget==3.2
